Allow installer to set slot user (#19)
* Use single slot with a slot user
* Allow install script to set slot user

The install script now changes the execute node config from
running jobs as nobody to running jobs as whoever is set as
slot user in the script. By default, this is the user running
the script.

The data source directory is now either:
1. Set to others-readonly permission if slotuser is nobody, or
2. Set to slotuser-readwrite permission if user is root, or
3. Set to xferusers group-readwrite permission and user and
   slotuser added to xferusers group if user is not slotuser, or
4. Left with default permissions if user is slotuser.

The result of all the above logic being that the slotuser will
have readwrite permission to the data source directory unless
the slotuser is set to nobody.