Adjusted install script to allow choice of docker or native
The Bash code is writted to work on both Linux and macOS,
and a check is done to force macOS users to use Docker.

Depending on if a user chooses docker or native exec modes,
either type of install is performed. For docker, the install
consists of creating a persistent ~/.htphenotyping directory
to store config.d and tokens.d, as well as a pointer to the
data source directory. The install script then calls run_docker.sh
to run the user through token generation and a (presumably) a
test run of HTCondor. Once the install script has run successfully,
the user only needs to execute run_docker.sh (no arguments needed)
to (re)start HTCondor since the config and token are persistent on
disk.

Note that the Docker image we are using is EL7-based, so the SSL
fix for Debian-based installs is problematic. We only add them now
for native installs.